requirements.txt
python-twitter
Pillow
pycoingecko
Google-Images-Search
schedule